move textParser to a separate file
